Create README.md
I have done a new layout of the PiHPSDR Controller V.2. The original controller use a 7 inch Touch Display (the original one made for Raspberry Pi). The new 5 inch version use a KeDei 5 inch 800x480 display with a capacitive touch screen and use the same interface as the original (not HDMI). The size of the PCB is 220x115 mm and fits in a Hammond extruded enclosure type 1455Q2201BK.